CPU Test #1 (#30)

+++ src/Types.hpp
@@ -3,7 +3,7 @@
 #include <cstdint>
 
 using u8 = std::uint_least8_t;
-using u16 = std::uint_least8_t;
+using u16 = std::uint_least16_t;
 
 using s8 = std::int_least8_t;
 using s16 = std::int_least16_t;
